add group transactions by day
