+++ src/components/Transactions.tsx
@@ -16,14 +16,14 @@ import {
   TableHeader,
   TableRow,
 } from "@/src/components/ui/table";
-import { TransactionsByMonthI } from "@/src/types";
+import { TransactionsGroupI } from "@/src/types";
 import Amount from "@/src/components/Amount";
 import { format } from "date-fns";
 
 export default function Transactions({
   transactions,
 }: {
-  transactions: TransactionsByMonthI[];
+  transactions: TransactionsGroupI[];
 }) {
   const maxVisibleTransactions = 50;
   const collapsedTransactions = transactions.slice(0, maxVisibleTransactions);
