+++ src/types.ts
@@ -56,7 +56,7 @@ export enum Institution {
   DKB = "DKB",
 }
 
-export interface TransactionsByMonthI {
+export interface TransactionsGroupI {
   group: string;
   date: string;
   transactions: TransactionI[];
@@ -65,5 +65,5 @@ export interface StatsI {
   totalBalance: number;
   expenses: number;
   income: number;
-  transactionsByMonth: TransactionsByMonthI[];
+  transactionsGroup: TransactionsGroupI[];
 }
